pyproject.toml
[tool.poetry]
name = "saturnsync"
version = "0.1.0"
description = ""
authors = ["Parker Wahle <parkeredwardwahle2017@gmail.com>"]
license = "GPLv3"
readme = "README.md"
homepage = "https://github.com/regulad/saturnsync"
repository = "https://github.com/regulad/saturnsync"

[tool.poetry.dependencies]
python = "^3.10"
saturnscrape = "^0.1.8"
aiohttp = "^3.8.1"
ics = "^0.7"
motor = "^2.5.1"
dnspython = "^2.1.0"
PyJWT = { extras = ["crypto"], version = "^2.3.0" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
